refactor naming, disable buttons
