+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityCreateModal/EntityCreateModalStyles.tsx
